+++ dotnet/src/Logging/Logging/SmartStandards/SmartStandardsTraceLogPipe.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Diagnostics;
-using System.Linq;
 
 namespace Logging.SmartStandards {
 
@@ -15,29 +14,37 @@ namespace Logging.SmartStandards {
   ///   2. If you copy and paste this class (instead of referencing this library) make sure you change namespace and class name. 
   ///      Otherwise you'll run into naming collisionsif a 3rd party DLL references this library.
   /// </remarks>
-  public class SmartStandardsTraceLogPipe : TraceListener {
+  public sealed class SmartStandardsTraceLogPipe : TraceListener {
 
     private static bool _Semaphore;
 
-    private Action<string, int, int, string, string[]> _OnLog;
+    private Action<string, int, int, string, object[]> _OnLog;
 
     private SmartStandardsTraceLogPipe() { // Constructor is private, because this must be a single instance
     }
 
+    /// <summary>
+    ///   Registers the pipe as trace listener, allowing trace sources to wire up to the pipe.
+    ///   Wires up your actual logging sink by registering a callback delegate.
+    /// </summary>
     /// <param name="onLog">
-    ///   Callback delegate which is called for each recieved log entry. Implement your actual logging code here.
-    ///   Make sure this is done before initializing the TraceSource.
+    ///   Callback delegate which is called for each recieved log entry. 
+    ///   Signature: onLog(string channelName, int level, int id, string messageTemplate, object[] args)
+    ///   Implement your actual logging sink here.
     /// </param>
-    public static void Initialize(Action<string, int, int, string, string[]> onLog) {
+    /// <remarks>
+    ///   Order dependency: This has to be done before initializing any trace sources 
+    ///   (= before any usage of a logger, because the loggers implicitely wire up their internal trace source on first usage).
+    /// </remarks>
+    public static void Initialize(Action<string, int, int, string, object[]> onLog) {
 
-      if (_Semaphore)
-        throw new Exception("Do not initialize more than once!");
+      if (_Semaphore) throw new Exception("Do not initialize more than once!");
 
-      var dummy = new SmartStandardsTraceLogPipe();
+      SmartStandardsTraceLogPipe instance = new SmartStandardsTraceLogPipe {
+        _OnLog = onLog
+      };
 
-      dummy._OnLog = onLog;
-
-      Trace.Listeners.Add(dummy); // Self-register to .net runtime
+      Trace.Listeners.Add(instance); // Self-register to .net runtime
 
       // Remark: This is a one-way-ticket. Removing (even disposing) a listener is futile, because all TraceSources hold a reference.
       // You would have to iterate all existing TraceSources first an remove the listener there.
@@ -62,55 +69,24 @@ namespace Logging.SmartStandards {
 
       // Map EventType => LogLevel
 
-      int level = this.EventTypeToLevel(eventType);
+      int level = 0; // Default: "Trace" (aka "Verbose")
+
+      switch (eventType) {
+        case TraceEventType.Critical: level = 5; break; // (aka "Fatal")
+        case TraceEventType.Error: level = 4; break;
+        case TraceEventType.Warning: level = 3; break;
+        case TraceEventType.Information: level = 2; break;
+        case TraceEventType.Transfer: level = 1; break; // There is no "Debug" EventType => (mis)use something else        
+      }
 
       // Push log entry to callback method
 
-      string[] argsAsStrings = null;
-
-      int argsMaxIndex = -1;
-
-      if (args != null) argsMaxIndex = args.GetUpperBound(0);
-
-      if (argsMaxIndex >= 0) {
-        argsAsStrings = new string[argsMaxIndex + 1];
-        for (int i = 0; i <= argsMaxIndex; i++) argsAsStrings[i] = args[i].ToString();
-      }
-
-      _OnLog.Invoke(source, level, eventId, messageTemplate, argsAsStrings);
+      _OnLog.Invoke(source, level, eventId, messageTemplate, args);
     }
 
-    private int EventTypeToLevel(TraceEventType eventType) {
-      switch (eventType) {
-        case TraceEventType.Critical: { // (aka "Fatal")
-          return 5;
-        }
-        case TraceEventType.Error: {
-          return 4;
-        }
-        case TraceEventType.Warning: {
-          return 3;
-        }
-        case TraceEventType.Information: {
-          return 2;
-        }
-        case TraceEventType.Transfer: { // There is no "Debug" EventType => use something else
-          return 1; // "Debug"
-        }
+    public override void Write(string message) { } // Not needed
 
-        default: {
-          return 0; // "Trace" (aka "Verbose")
-        }
-      }
-    }
-
-    public override void Write(string message) {
-      // Not needed
-    }
-
-    public override void WriteLine(string message) {
-      // Not needed
-    }
+    public override void WriteLine(string message) { }// Not needed
 
   }