+++ dotnet/test/Logging.Tests/Logging/SmartStandards/SmartStandardsTraceLogPipeTests.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Logging.SmartStandards {
+
+  [TestClass()]
+  public class SmartStandardsTraceLogPipeTests : LoggerTestsBase {
+
+    internal override bool UseTrace { get { return true; } }
+
+  }
+}